Move new config behind feature toggle
